mise a jour des routes et finalisations de la documetation des nouvelles fonctionnalités

+++ app/Http/Controllers/EtudeCasController.php
@@ -17,7 +17,17 @@ use OpenApi\Annotations as OA;
 
 class EtudeCasController extends Controller
 {
-    //utiliser cette fonction si vous voulez  lister les etudes de cas
+     /**
+     * @OA\Get(
+     *     path="/api/index_cas",
+     *     summary="Lister tous les  cas de la plateforme",
+     *     tags={"EtudeCas"},
+     *     @OA\Response(
+     *         response=200,
+     *         description="Liste de tous les cas",
+     *     ),
+     * )
+     */
     public function index()
     {
         $etudeCas = EtudeCas::all();
@@ -26,7 +36,7 @@ class EtudeCasController extends Controller
 
     /**
      * @OA\Post(
-     *      path="/api/etude-cas/create/{id}",
+     *      path="/api/etude-cas/create",
      *      operationId="createEtudeCas",
      *      tags={"EtudeCas"},
      *      summary="Ajouter une nouvelle étude de cas",
@@ -51,41 +61,52 @@ class EtudeCasController extends Controller
      *          response=200,
      *          description="Etude de cas ajoutée avec succès"
      *      ),
-     *      security={
-     *          {"api_key": {}}
+     *       security={
+     *          {"Bearer": {}}
      *      }
      * )
      */
-    public function create_cas(Request $request)
+    public function create(Request $request)
     {
-
-        $imagePath = null;
-        if ($request->hasFile('image')) {
-            $image = $request->file('image');
-            $imageName = time() . '.' . $image->getClientOriginalExtension();
-            $imagePath = $image->storeAs('images', $imageName, 'public');
+        if (!auth()->check()) {
+            return response()->json(['message' => 'Non autorisé'], 401);
         }
-
-        $etudeCas = $request->validate([
-
-            'contenu' => 'required',
-            'image' => 'required',
-
-        ]);
-        $etudeCas = new EtudeCas($etudeCas);
-        $etudeCas->contenu = $request->contenu;
-        $etudeCas->image = $imagePath;
-        $etudeCas->user_id = FacadesAuth::user()->id;
-        $etudeCas->secteur_id = $request->secteur_id;
-        $etudeCas->save();
-
-        return response()->json(['message' => 'etudeCas ajouter avec succée', 'etudeCas' => $etudeCas], 200);
-    }
-
+        $user=auth()->user();
+        if ($user->role->id===2) {  
+            $imagePath = null;
+            if ($request->hasFile('image')) {
+                $image = $request->file('image');
+                $imageName = time() . '.' . $image->getClientOriginalExtension();
+                $imagePath = $image->storeAs('images', $imageName, 'public');
+            }
+    
+            $etudeCas = $request->validate([
+    
+                'contenu' => 'required',
+                'image' => 'required',
+    
+            ]);
+            $etudeCas = new EtudeCas($etudeCas);
+            $etudeCas->contenu = $request->contenu;
+            $etudeCas->image = $imagePath;
+            $etudeCas->user_id = FacadesAuth::user()->id;
+            $etudeCas->secteur_id = $request->secteur_id;
+            $etudeCas->save();
+    
+            return response()->json(['message' => 'etudeCas ajouter avec succée', 'etudeCas' => $etudeCas], 200);
+        }
+        
+            return response()->json(
+                [
+                    "message"=>"Vous n'avez pas les droits pour ajouter une étude de cas",
+                    403
+                ],
+            );
+        }
 
     /**
      * @OA\Post(
-     *      path="/api/etude-cas/update_etudeCas",
+     *      path="/api/update_etudeCas",
      *      operationId="updateEtudeCas",
      *      tags={"EtudeCas"},
      *      summary="Modifier une étude de cas existante",
@@ -105,20 +126,24 @@ class EtudeCasController extends Controller
      *          description="Etude de cas modifiée avec succès")
      *      ),
      *      security={
-     *          {"api_key": {}}
+     *          {"Bearer": {}}
      *      }
      * )
      */
     public function update(Request $request)
     {
+        if (!auth()->check()) {
+            return response()->json(['message' => 'Non autorisé'], 401);
+        }
+       
+        $etudeCas = EtudeCas::find($request->id);
+        if (auth()->user()->id===$etudeCas->user_id) {
         $imagePath = null;
         if ($request->hasFile('image')) {
             $image = $request->file('image');
             $imageName = time() . '.' . $image->getClientOriginalExtension();
             $imagePath = $image->storeAs('images', $imageName, 'public');
         }
-
-        $etudeCas = EtudeCas::find($request->id);
         $etudeCas->contenu = $request->contenu;
         $etudeCas->image = $imagePath;
         $etudeCas->user_id = FacadesAuth::user()->id;
@@ -127,6 +152,11 @@ class EtudeCasController extends Controller
 
         return response()->json(['message' => 'etude cas modifiée avec succès', 'etudecas' => $etudeCas], 200);
     }
+        return response()->json(['message' => 'Vous n\'avez pas le droit
+        de modifier cette étude de cas'],
+         401);
+
+}
 
     /**
      * @OA\Post(
@@ -146,18 +176,23 @@ class EtudeCasController extends Controller
      *          response=200,
      *          description="Etude de cas archivée avec succès",
      *      ),
-     *      security={
-     *          {"api_key": {}}
+     *       security={
+     *          {"Bearer": {}}
      *      }
      * )
      */
     public function archive(Request $request)
     {
+        if (!auth()->check()) {
+            return response()->json(['message' => 'Non autorisé'], 401);
+        }
         $etudeCas = EtudeCas::find($request->id);
         $etudeCas->is_deleted = true;
         $etudeCas->save();
 
-        return response()->json(['message' => 'etude cas archivée avec succès', 'etudecas' => $etudeCas], 200);
+        return response()->json([
+        'message' => 'etude cas archivée avec succès',
+         'etudecas' => $etudeCas], 200);
     }
 
     /**
@@ -179,12 +214,15 @@ class EtudeCasController extends Controller
      *          description="Etude de cas supprimée avec succès",
      *      ),
      *      security={
-     *          {"api_key": {}}
+     *          {"Bearer": {}}
      *      }
      * )
      */
     public function delete(Request $request)
     {
+        if (!auth()->check()) {
+            return response()->json(['message' => 'Non autorisé'], 401);
+        }
         $etudeCas = EtudeCas::find($request->id);
         $etudeCas->delete();
